pyproject.toml
[project]
name = "cicd"
version = "0.1.0"
description = ""
authors = [
    {name = "stavros-vlach",email = "stavrosvlach34@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-functions (>=1.23.0,<2.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "pytest (>=8.4.2,<9.0.0)"
]

[tool.poetry]
packages = [{include = "cicd", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
